+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+- Version command `dinghy -v`.
+
 ## 2.0.2 - 2014-12-05
 
 ### Changed